fix so that all pages have the same width and continue with post ad page
