+++ src/app/(with-nav)/product/components/Navigation.tsx
@@ -1,0 +1,14 @@
+import Link from "next/link";
+import Image from "next/image";
+
+export default function Navigation(props: { readonly name: string }) {
+  return (
+    <div className="mt-3 flex space-x-2">
+      <Link className="text-sm hover:underline" href="/">
+        Start
+      </Link>
+      <Image src="/images/arrow.svg" alt="arrow" width="8" height="8" />
+      <p className="text-sm text-gray-500">{props.name}</p>
+    </div>
+  );
+}
