+++ src/app/(with-nav)/profile/page.tsx
@@ -4,7 +4,7 @@ import MyMarketplace from "@/app/(with-nav)/profile/components/MyMarketplace";
 
 export default function Profile() {
   return (
-    <div className="mx-auto mt-5 w-8/12 px-2">
+    <div className="mx-auto mt-5 w-8/12">
       <MyProfile />
       <MyMarketplace />
       <Logout />
