+++ src/app/(with-nav)/profile/purchases/page.tsx
@@ -3,7 +3,7 @@ import Navigation from "@/app/(with-nav)/profile/watchlist/components/Navigation
 
 export default function Purchases() {
   return (
-    <div className="mx-auto mt-2 w-8/12 px-2">
+    <div className="mx-auto mt-2 w-8/12">
       <Navigation name="Purchases" />
       <div className="mt-8">
         <MyPurchases />
