+++ src/app/(with-nav)/profile/watchlist/page.tsx
@@ -3,7 +3,7 @@ import Navigation from "@/app/(with-nav)/profile/watchlist/components/Navigation
 
 export default function Inbox() {
   return (
-    <div className="mx-auto mt-3 w-8/12 px-2">
+    <div className="mx-auto mt-3 w-8/12">
       <Navigation name="Watchlist" />
       <MyWatchlist />
     </div>
